[draw] Move draw_extents form paint-extents to draw
